+++ src/Application/Query/Meeting/ParticipantsViewQuery.php
@@ -12,6 +12,8 @@ namespace Proximum\Vimeet\Application\Query\Meeting;
 
 use Proximum\Vimeet\Domain\Model\Contact;
 use Proximum\Vimeet\Domain\Model\Participant;
+use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Domain\Model\User;
 
 class ParticipantsViewQuery
 {
@@ -24,17 +26,22 @@ class ParticipantsViewQuery
     /** @var Contact[] */
     public $contacts;
 
+    /** @var Sheet */
+    public $seerSheet;
+
+    /** @var User */
+    public $seerUser;
+
     /**
-     * ParticipantsViewQuery constructor.
-     *
      * @param Participant[] $participants
-     * @param string        $locale
      * @param Contact[]     $contacts
      */
-    public function __construct(array $participants, $locale, array $contacts)
+    public function __construct(array $participants, string $locale, array $contacts, Sheet $seerSheet, User $seerUser)
     {
         $this->participants = $participants;
         $this->locale = $locale;
         $this->contacts = $contacts;
+        $this->seerSheet = $seerSheet;
+        $this->seerUser = $seerUser;
     }
 }